requirements.txt
jinja2==2.11.1
pyyaml==5.3.1
click==7.1.2
cli_helpers[styles] >= 2.0.0
psycopg2-binary >= 2.8
